removed some logic from payload_request class, namely the record_payload method
