+++ test/controllers/receive_client_payload_test.rb
@@ -28,7 +28,7 @@ class ReceiveClientPayloadTest < Minitest::Test
     Client.create(identifier: "Client1", root_url: "www.client.com")
 
     payload = '{"url":"http://jumpstartlab.com/blog","requestedAt":"2013-02-16 21:38:28 -0700","respondedIn":37,"referredBy":"http://jumpstartlab.com","requestType":"GET","parameters":[],"eventName":"socialLogin","userAgent":"Mozilla/5.0 (Macintosh%3B Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17","resolutionWidth":"1920","resolutionHeight":"1280","ip":"63.29.38.211"}'
-    PayloadRequest.record_payload(payload, "Client1")
+    PayloadCreator.record_payload(payload, "Client1")
     assert_equal 1, PayloadRequest.count
 
     post '/sources/Client1/data', {payload: payload}
