+++ test/models/payload_request_test.rb
@@ -7,7 +7,7 @@ class PayloadRequestTest < Minitest::Test
   def test_can_write_raw_json_to_database
     raw_json = '{"url":"http://jumpstartlab.com/blog","requestedAt":"2013-02-16 21:38:28 -0700","respondedIn":37,"referredBy":"http://jumpstartlab.com","requestType":"GET","parameters":[],"eventName": "socialLogin","userAgent":"Mozilla/5.0 (Macintosh%3B Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17","resolutionWidth":"1920","resolutionHeight":"1280","ip":"63.29.38.211"}'
     Client.create(identifier: "client1", root_url: "www.example.com")
-    PayloadRequest.record_payload(raw_json, "client1")
+    PayloadCreator.record_payload(raw_json, "client1")
 
     assert_equal 1, PayloadRequest.count
 
@@ -31,10 +31,10 @@ class PayloadRequestTest < Minitest::Test
   def test_it_handles_similar_payload_requests
     raw_json_1 = '{"url":"http://jumpstartlab.com/blog","requestedAt":"2013-02-16 21:38:28 -0700","respondedIn":37,"referredBy":"http://jumpstartlab.com","requestType":"GET","parameters":[],"eventName": "socialLogin","userAgent":"Mozilla/5.0 (Macintosh%3B Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17","resolutionWidth":"1920","resolutionHeight":"1280","ip":"63.29.38.211"}'
     Client.create(identifier: "client1", root_url: "www.example.com")
-    PayloadRequest.record_payload(raw_json_1, "client1")
+    PayloadCreator.record_payload(raw_json_1, "client1")
 
     raw_json_2 = '{"url":"http://jumpstartlab.com/blog","requestedAt":"2014-02-16 21:38:28 -0700","respondedIn":37,"referredBy":"http://jumpstartlab.com","requestType":"GET","parameters":[],"eventName": "socialLogin","userAgent":"Mozilla/5.0 (Macintosh%3B Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17","resolutionWidth":"1920","resolutionHeight":"1280","ip":"63.29.38.211"}'
-    PayloadRequest.record_payload(raw_json_2, "client1")
+    PayloadCreator.record_payload(raw_json_2, "client1")
 
     assert_equal 2, PayloadRequest.count
     assert_equal 1, Referrer.count
@@ -262,7 +262,7 @@ class PayloadRequestTest < Minitest::Test
     client_identifier = "BestBuy"
     raw_payload = '{"url":"http://jumpstartlab.com/blog","requestedAt":"2013-02-16 21:38:28 -0700","respondedIn":37,"referredBy":"http://jumpstartlab.com","requestType":"GET","parameters":[],"eventName": "socialLogin","userAgent":"Mozilla/5.0 (Macintosh%3B Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17","resolutionWidth":"1920","resolutionHeight":"1280","ip":"63.29.38.211"}'
 
-    PayloadRequest.record_payload(raw_payload, client_identifier)
+    PayloadCreator.record_payload(raw_payload, client_identifier)
     assert_equal 1, PayloadRequest.count
 
     boolean = PayloadRequest.duplicate?(raw_payload, client_identifier)
